Fixed memory leak (forgot to change free size of chunk from 8 to 64 bits), working on vm main loop

+++ src/lit/vm/lit_chunk.c
@@ -16,7 +16,7 @@ void lit_init_chunk(LitChunk* chunk) {
 }
 
 void lit_free_chunk(LitState* state, LitChunk* chunk) {
-	LIT_FREE_ARRAY(state, uint8_t, chunk->code, chunk->capacity);
+	LIT_FREE_ARRAY(state, uint64_t, chunk->code, chunk->capacity);
 	LIT_FREE_ARRAY(state, uint16_t , chunk->lines, chunk->line_capacity);
 
 	lit_free_values(state, &chunk->constants);
